Remove async calls by pregenerating vk tree instead of computing on the fly
