+++ yarn-project/.gitignore
@@ -41,6 +41,7 @@ noir-protocol-circuits-types/src/private_kernel_reset_data.ts
 noir-protocol-circuits-types/src/private_kernel_reset_vks.ts
 noir-protocol-circuits-types/src/private_kernel_reset_types.ts
 noir-protocol-circuits-types/src/client_artifacts_helper.ts
+noir-protocol-circuits-types/src/vk_tree.ts
 noir-protocol-circuits-types/src/types/
 ivc-integration/artifacts
 ivc-integration/src/types/
