+++ yarn-project/noir-protocol-circuits-types/src/entrypoint/vks.ts
@@ -1,126 +1,19 @@
-import {
-  BASE_PARITY_INDEX,
-  BLOCK_MERGE_ROLLUP_INDEX,
-  BLOCK_ROOT_ROLLUP_EMPTY_INDEX,
-  BLOCK_ROOT_ROLLUP_INDEX,
-  BLOCK_ROOT_ROLLUP_SINGLE_TX_INDEX,
-  Fr,
-  MERGE_ROLLUP_INDEX,
-  type MerkleTree,
-  MerkleTreeCalculator,
-  PRIVATE_BASE_ROLLUP_VK_INDEX,
-  PRIVATE_KERNEL_INIT_INDEX,
-  PRIVATE_KERNEL_INNER_INDEX,
-  PRIVATE_KERNEL_TAIL_INDEX,
-  PRIVATE_KERNEL_TAIL_TO_PUBLIC_INDEX,
-  PUBLIC_BASE_ROLLUP_VK_INDEX,
-  ROOT_PARITY_INDEX,
-  ROOT_ROLLUP_INDEX,
-  TUBE_VK_INDEX,
-  VK_TREE_HEIGHT,
-  VerificationKeyAsFields,
-  VerificationKeyData,
-} from '@aztec/circuits.js';
-import { poseidon2Hash } from '@aztec/foundation/crypto';
+import { Fr, VK_TREE_HEIGHT, VerificationKeyAsFields, VerificationKeyData } from '@aztec/circuits.js';
 import { assertLength } from '@aztec/foundation/serialize';
 
-import BaseParityVkJson from '../../artifacts/keys/parity_base.vk.data.json' assert { type: 'json' };
-import RootParityVkJson from '../../artifacts/keys/parity_root.vk.data.json' assert { type: 'json' };
-import PrivateKernelInitVkJson from '../../artifacts/keys/private_kernel_init.vk.data.json' assert { type: 'json' };
-import PrivateKernelInnerVkJson from '../../artifacts/keys/private_kernel_inner.vk.data.json' assert { type: 'json' };
-import PrivateKernelTailVkJson from '../../artifacts/keys/private_kernel_tail.vk.data.json' assert { type: 'json' };
-import PrivateKernelTailToPublicVkJson from '../../artifacts/keys/private_kernel_tail_to_public.vk.data.json' assert { type: 'json' };
-import PrivateBaseRollupVkJson from '../../artifacts/keys/rollup_base_private.vk.data.json' assert { type: 'json' };
-import PublicBaseRollupVkJson from '../../artifacts/keys/rollup_base_public.vk.data.json' assert { type: 'json' };
-import BlockMergeRollupVkJson from '../../artifacts/keys/rollup_block_merge.vk.data.json' assert { type: 'json' };
-import BlockRootRollupVkJson from '../../artifacts/keys/rollup_block_root.vk.data.json' assert { type: 'json' };
-import EmptyBlockRootRollupVkJson from '../../artifacts/keys/rollup_block_root_empty.vk.data.json' assert { type: 'json' };
-import SingleTxBlockRootRollupVkJson from '../../artifacts/keys/rollup_block_root_single_tx.vk.data.json' assert { type: 'json' };
-import MergeRollupVkJson from '../../artifacts/keys/rollup_merge.vk.data.json' assert { type: 'json' };
-import RootRollupVkJson from '../../artifacts/keys/rollup_root.vk.data.json' assert { type: 'json' };
-import TubeVkJson from '../../artifacts/keys/tube.vk.data.json' assert { type: 'json' };
-import { type ClientProtocolArtifact, type ProtocolArtifact, type ServerProtocolArtifact } from '../artifacts/types.js';
-import { PrivateKernelResetVkIndexes, PrivateKernelResetVks } from '../private_kernel_reset_vks.js';
-import { keyJsonToVKData } from '../utils/vk_json.js';
+import { vkTree } from '../vk_tree.js';
 
-// TODO Include this in the normal maps when the tube is implemented in noir
-export const TubeVk = keyJsonToVKData(TubeVkJson);
+export * from '../artifacts/vks.js';
 
-export const ServerCircuitVks: Record<ServerProtocolArtifact, VerificationKeyData> = {
-  BaseParityArtifact: keyJsonToVKData(BaseParityVkJson),
-  RootParityArtifact: keyJsonToVKData(RootParityVkJson),
-  PrivateBaseRollupArtifact: keyJsonToVKData(PrivateBaseRollupVkJson),
-  PublicBaseRollupArtifact: keyJsonToVKData(PublicBaseRollupVkJson),
-  MergeRollupArtifact: keyJsonToVKData(MergeRollupVkJson),
-  BlockRootRollupArtifact: keyJsonToVKData(BlockRootRollupVkJson),
-  SingleTxBlockRootRollupArtifact: keyJsonToVKData(SingleTxBlockRootRollupVkJson),
-  EmptyBlockRootRollupArtifact: keyJsonToVKData(EmptyBlockRootRollupVkJson),
-  BlockMergeRollupArtifact: keyJsonToVKData(BlockMergeRollupVkJson),
-  RootRollupArtifact: keyJsonToVKData(RootRollupVkJson),
-};
-
-export const ClientCircuitVks: Record<ClientProtocolArtifact, VerificationKeyData> = {
-  PrivateKernelInitArtifact: keyJsonToVKData(PrivateKernelInitVkJson),
-  PrivateKernelInnerArtifact: keyJsonToVKData(PrivateKernelInnerVkJson),
-  PrivateKernelTailArtifact: keyJsonToVKData(PrivateKernelTailVkJson),
-  PrivateKernelTailToPublicArtifact: keyJsonToVKData(PrivateKernelTailToPublicVkJson),
-  ...PrivateKernelResetVks,
-};
-
-export const ProtocolCircuitVks: Record<ProtocolArtifact, VerificationKeyData> = {
-  ...ClientCircuitVks,
-  ...ServerCircuitVks,
-};
-
-export const ProtocolCircuitVkIndexes: Record<ProtocolArtifact, number> = {
-  PrivateKernelInitArtifact: PRIVATE_KERNEL_INIT_INDEX,
-  PrivateKernelInnerArtifact: PRIVATE_KERNEL_INNER_INDEX,
-  PrivateKernelTailArtifact: PRIVATE_KERNEL_TAIL_INDEX,
-  PrivateKernelTailToPublicArtifact: PRIVATE_KERNEL_TAIL_TO_PUBLIC_INDEX,
-  BaseParityArtifact: BASE_PARITY_INDEX,
-  RootParityArtifact: ROOT_PARITY_INDEX,
-  PrivateBaseRollupArtifact: PRIVATE_BASE_ROLLUP_VK_INDEX,
-  PublicBaseRollupArtifact: PUBLIC_BASE_ROLLUP_VK_INDEX,
-  MergeRollupArtifact: MERGE_ROLLUP_INDEX,
-  BlockRootRollupArtifact: BLOCK_ROOT_ROLLUP_INDEX,
-  SingleTxBlockRootRollupArtifact: BLOCK_ROOT_ROLLUP_SINGLE_TX_INDEX,
-  EmptyBlockRootRollupArtifact: BLOCK_ROOT_ROLLUP_EMPTY_INDEX,
-  BlockMergeRollupArtifact: BLOCK_MERGE_ROLLUP_INDEX,
-  RootRollupArtifact: ROOT_ROLLUP_INDEX,
-  ...PrivateKernelResetVkIndexes,
-};
-
-async function buildVKTree() {
-  const calculator = await MerkleTreeCalculator.create(VK_TREE_HEIGHT, Buffer.alloc(32), async (a, b) =>
-    (await poseidon2Hash([a, b])).toBuffer(),
-  );
-  const vkHashes = new Array(2 ** VK_TREE_HEIGHT).fill(Buffer.alloc(32));
-
-  for (const [key, value] of Object.entries(ProtocolCircuitVks)) {
-    const index = ProtocolCircuitVkIndexes[key as ProtocolArtifact];
-    vkHashes[index] = value.keyAsFields.hash.toBuffer();
-  }
-
-  vkHashes[TUBE_VK_INDEX] = TubeVk.keyAsFields.hash.toBuffer();
-
-  return calculator.computeTree(vkHashes);
-}
-
-let vkTree: MerkleTree | undefined;
-
-export async function getVKTree() {
-  if (!vkTree) {
-    vkTree = await buildVKTree();
-  }
+export function getVKTree() {
   return vkTree;
 }
 
-export async function getVKTreeRoot() {
-  const tree = await getVKTree();
-  return Fr.fromBuffer(tree.root);
+export function getVKTreeRoot() {
+  return Fr.fromBuffer(vkTree.root);
 }
 
-export async function getVKIndex(vk: VerificationKeyData | VerificationKeyAsFields | Fr) {
+export function getVKIndex(vk: VerificationKeyData | VerificationKeyAsFields | Fr) {
   let hash;
   if (vk instanceof VerificationKeyData) {
     hash = vk.keyAsFields.hash;
@@ -130,19 +23,16 @@ export async function getVKIndex(vk: VerificationKeyData | VerificationKeyAsFiel
     hash = vk;
   }
 
-  const tree = await getVKTree();
-
-  const index = tree.getIndex(hash.toBuffer());
+  const index = vkTree.getIndex(hash.toBuffer());
   if (index < 0) {
     throw new Error(`VK index for ${hash.toString()} not found in VK tree`);
   }
   return index;
 }
 
-export async function getVKSiblingPath(vkIndex: number) {
-  const tree = await getVKTree();
+export function getVKSiblingPath(vkIndex: number) {
   return assertLength<Fr, typeof VK_TREE_HEIGHT>(
-    tree.getSiblingPath(vkIndex).map(buf => new Fr(buf)),
+    vkTree.getSiblingPath(vkIndex).map(buf => new Fr(buf)),
     VK_TREE_HEIGHT,
   );
 }
