+++ yarn-project/prover-client/src/orchestrator/orchestrator_multi_public_functions.test.ts
@@ -36,7 +36,7 @@ describe('prover/orchestrator/public-functions', () => {
         );
         for (const tx of txs) {
           tx.data.constants.historicalHeader = context.getBlockHeader(0);
-          tx.data.constants.vkTreeRoot = await getVKTreeRoot();
+          tx.data.constants.vkTreeRoot = getVKTreeRoot();
           tx.data.constants.protocolContractTreeRoot = protocolContractTreeRoot;
         }
