+++ yarn-project/pxe/src/kernel_prover/kernel_prover.ts
@@ -214,7 +214,7 @@ export class KernelProver {
       if (firstIteration) {
         const proofInput = new PrivateKernelInitCircuitPrivateInputs(
           txRequest,
-          await getVKTreeRoot(),
+          getVKTreeRoot(),
           protocolContractTreeRoot,
           privateCallData,
           isPrivateOnlyTx,
